Move ssh runner stuff to st2actions/runners/ssh
